backend/requirements.txt
Flask
cryptography
python-dotenv
PyJWT
web3